+++ HISTORY.rst
@@ -4,7 +4,7 @@
 History
 -------
 
-5.3.0
+5.3.0 (2026-07-16)
 ++++++++++++++++++
 
 * The version is now retrieved from package metadata at runtime using